requirements.txt
streamlit
pandas
fredapi
yfinance
plotly
python-dotenv
